+++ src/utils/__tests__/elasticsearch2.spec.js
